Add a github action to trigger binder build (#3)
When the repository is updated, the first user who clicks the "launch binder" badge has to wait ~10 minutes for the building of the docker images. After completing the building of docker images, other users only need ~15 seconds to launch the Jupiter lab. Thus, the first user will have a terrible experience.

To resolve the problem, we need to manually trigger binder builds after the master branch changes. This can be done by using the Binder build API. This github action (https://github.com/marketplace/actions/trigger-binder-build) can trigger binder builds automatically on 5 mybinder services but doesn't support the pangeo binder.

This PR adds a github action which triggers binder build for the following 6 binder services:

https://binder.pangeo.io/
https://mybinder.org/ (this one is always redirected to the 4 subdomains below)
https://gke.mybinder.org/
https://ovh.mybinder.org/
https://gesis.mybinder.org/
https://turing.mybinder.org/

The links of the 6 binder services are all listed in the README, so that users can have more choices if one is not responding.